Allows the first argument to to_struct/2 to be a struct
This seems weird at first, but can be useful in certain applications. If
for example you define a domain struct, that gets passed around your
application, which gets turned into an Ecto struct just before we
insert/update etc in the DB. Often all the fields are the same, we just
want to convert between the two.