Fixes a bug with to_struct
When an embedded changeset is invalid, but the parent is not invalid,
the parent changeset is marked as valid?: false by ecto, but the errors
contained in the child changeset are not present in the errors key on
the parent changeset. We were assuming that changesets were valid if the
errors key was an empty list in ecto, but instead we should be using the
valid? key to determine that.

In short changesets can look like this:

Ecto.Changeset{
   changes: %{
     child: Ecto.Changeset{
       action: :insert,
       changes: %{location: "Kitchen"},
       errors: [
         actually_a_fire?: {"is invalid", [type: :boolean, validation: :cast]}
       ],
       data: %{},
       valid?: false
     },
   },
   data: %{},
   errors: [],
   valid?: false
 }