Possibly working version...
Obvs needs propert testing now. But this is in danger of allowing
keyword lens style list of fields to require. Allows the fields to be
relations and even allows them to have deeper relations on them.

Lots of dimensions to testing though:

has / embeds one / many
not in data / not in changes
path to nested changeset comes up short (doesn't point to a thing), for
changesets we want to be able to not raise in that case, and just add an
error to the changeset I guess...

There is more...